Refactor ns events
