+++ code/cell.typesystem/src/TypeSystem.core/TypeCache/index.ts
@@ -1,1 +1,2 @@
 export * from './TypeCache';
+export * from './TypeCacheKey';
